automation-chatbot-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.0.0,<2.0.0
anthropic>=0.39.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
supabase>=2.0.0
jsonschema>=4.20.0
pyyaml>=6.0.1
python-multipart>=0.0.6
httpx>=0.25.0
pytest>=7.4.3
pytest-asyncio>=0.21.1

# AI service dependencies
tenacity>=8.2.3
tiktoken>=0.5.2
aiofiles>=23.2.1

# Payment processing
stripe>=7.0.0

# S3-compatible storage for data archiving
boto3>=1.34.0
